Removes copy button and adds blank slate fill
When a machine matches the streamline inventory it now displays a copy
button with that machine’s info and user only has to click on that
button to copy.

If the form’s stock number is blank, a button to add from inventory is
displayed and when clicked shows the whole inventory in a sidebar. When
the user clicks a machine that info is sent to the form.